+++ libs/user/src/lib/user/+state/actions/frnds_new_user.actions.ts
@@ -1,7 +1,7 @@
 import { HttpErrorResponse } from '@angular/common/http';
 import { createAction, props } from '@ngrx/store';
 
-import { FrndsAppDetailActionTypes, FrndsAppNewUserActionTypes } from '../../types/actions/friends-app-action.types';
+import { FrndsAppDetailActionTypes, FrndsAppNewUserActionTypes } from '../../types/friends-app-action.types';
 import { User } from '../../types/frnds-app-state.interface';
 
 export const frndsAppNewUserClickAction = createAction(
